Restructures base app directory structure
