+++ index.html
@@ -95,7 +95,7 @@
     <!-- Load webcomponents-loader.js to check and load any polyfills your browser needs -->
     <script src="node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
     <!-- Built with love using PWA Starter Kit -->
-    <script type="module" src="src/components/marsfarm-app.js" crossorigin></script>
+    <script type="module" src="src/marsfarm-app.js" crossorigin></script>
     <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
   </body>
 </html>
